Refactoring/componentalization + Menu edit option (choicecard)
-Refactoring/component on frontend
-also added the 4th option to edit/add new menu
-changed spacing for analytics
-changed graphs using shadcn
