+++ src/app/inventory/types.ts
@@ -1,0 +1,10 @@
+// src/app/inventory/types.ts
+
+export interface InventoryItem {
+  id: number;
+  product: string;
+  category: string;
+  stock: number;
+  status: string;
+  cost: string;
+}
