+++ src/app/menu/types.ts
@@ -1,0 +1,10 @@
+// src/app/menu/types.ts
+
+// Using the same structure as the POS product model
+export interface MenuItem {
+  id: number;
+  name: string;
+  price: number;
+  category: string;
+  image: string; // Placeholder for future use
+}
