+++ src/app/order/components/OrderHeader.tsx
@@ -1,0 +1,74 @@
+// src/app/order/components/OrderHeader.tsx
+import React from 'react';
+// IMPORT FIXED: Added CoffeeTogo icon and useRouter hook
+import { Coffee, ClipboardPen, PieChart, Boxes, Menu } from 'lucide-react';
+import DropdownItem from '@/components/SelectionMenu/DropdownItem'; 
+import { useRouter } from 'next/navigation'; // Necessary to handle the new route
+
+interface OrderHeaderProps {
+  formattedTime: string;
+  isDropdownOpen: boolean;
+  setIsDropdownOpen: (isOpen: boolean) => void;
+  handleLogoClick: () => void;
+  handleOrderClick: () => void;
+  handleAnalyticsClick: () => void;
+  handleInventoryClick: () => void;
+}
+
+export default function OrderHeader({
+  formattedTime,
+  isDropdownOpen,
+  setIsDropdownOpen,
+  handleLogoClick,
+  handleOrderClick,
+  handleAnalyticsClick,
+  handleInventoryClick,
+}: OrderHeaderProps) {
+  const router = useRouter(); // Initialize router
+  const handleMenuClick = () => router.push('/menu'); // New handler for Menu link
+
+  return (
+    <header className="flex w-full items-center justify-between flex-shrink-0 relative z-30">
+      <div
+        className="relative z-50"
+        onMouseEnter={() => setIsDropdownOpen(true)}
+        onMouseLeave={() => setIsDropdownOpen(false)}
+      >
+        <div
+          className="flex cursor-pointer items-center gap-4 transition-opacity hover:opacity-80 pb-1"
+          onClick={handleLogoClick}
+        >
+          <div className="drop-shadow-md">
+            <Coffee size={72} className="text-gray-900" />
+          </div>
+          <span className="text-[64px] font-black leading-tight text-gray-900 drop-shadow-sm">
+            Pres <span className="text-[#6290C3]">Kopee</span>
+          </span>
+        </div>
+
+        <div
+          className={`
+            absolute left-0 top-full w-64 overflow-hidden rounded-xl bg-white shadow-xl ring-1 ring-black ring-opacity-5 transition-all duration-200 ease-in-out
+            ${isDropdownOpen ? 'max-h-96 opacity-100 translate-y-0' : 'max-h-0 opacity-0 -translate-y-2 pointer-events-none'}
+          `}
+        >
+          <div className="py-2">
+            <DropdownItem icon={ClipboardPen} label="Order" onClick={handleOrderClick} />
+            <DropdownItem icon={PieChart} label="Analytics" onClick={handleAnalyticsClick} />
+            <DropdownItem icon={Boxes} label="Inventory" onClick={handleInventoryClick} />
+            <DropdownItem icon={Menu} label="Menu" onClick={handleMenuClick} /> {/* NEW LINK ADDED */}
+          </div>
+        </div>
+      </div>
+
+      <div className="flex items-baseline gap-3 font-black italic tracking-tight drop-shadow-sm">
+        <span className="text-[64px] text-[#6290C3]">
+          {formattedTime.split(' ')[0]}
+        </span>
+        <span className="text-gray-900 text-[48px]">
+          {formattedTime.split(' ')[1]}
+        </span>
+      </div>
+    </header>
+  );
+}
