+++ src/app/order/components/ProductCard.tsx
@@ -1,0 +1,40 @@
+// src/app/order/components/ProductCard.tsx
+import React from 'react';
+import Image from 'next/image';
+import { Plus } from 'lucide-react';
+import { Product } from '../types';
+
+interface ProductCardProps {
+  product: Product;
+  onCustomize: () => void;
+}
+
+export default function ProductCard({ product, onCustomize }: ProductCardProps) {
+  return (
+    <div
+      className="
+        group flex flex-col justify-between items-center rounded-2xl bg-white p-4
+        shadow-md transition-all duration-200
+        hover:shadow-xl hover:-translate-y-0.5 border border-gray-100
+        cursor-pointer h-full min-h-[150px]
+      "
+      onClick={onCustomize}
+    >
+      {/* Remove Image component as it was just a placeholder and might cause build errors */}
+      {/* <Image src={product.image} alt={product.name} width={150} height={150} className="rounded-lg mb-2" /> */}
+      <div className="flex flex-col items-center justify-center flex-grow text-center">
+        <h3 className="text-xl font-extrabold text-gray-900 leading-snug">{product.name}</h3>
+        <p className="text-lg font-bold text-[#6290C3] mt-1">PHP {product.price.toFixed(2)}</p>
+      </div>
+
+      <div className="mt-4 w-full flex-shrink-0">
+        <div
+          className="w-full flex items-center justify-center gap-2 rounded-lg bg-[#E5F1FB] p-2 text-[#6290C3] font-semibold transition-all group-hover:bg-[#6290C3] group-hover:text-white"
+        >
+          <Plus size={18} />
+          <span>Select</span>
+        </div>
+      </div>
+    </div>
+  );
+}
